requirements.txt
selenium
pillow
pyTelegramBotAPI
